% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifai.R
\name{classifai}
\alias{classifai}
\title{Selecciona categorias para classificar los datos}
\usage{
classifai(
  dt,
  cat = NULL,
  ncat = "",
  sol_tem = NULL,
  api_key = NULL,
  temperature = 0.5,
  cont = "",
  cat_temperature = 0.5
)
}
\arguments{
\item{dt}{vector de elementos caracter}

\item{cat}{lista o vector con las categorias para classificar los datos, en el caso de lista ha de tener un element desc con la descripccion de las categorias y un elemento cat, vector con las categorias}

\item{ncat}{numeric}

\item{sol_tem}{elemento character para pedir a la funcion para que van a ser las categorias, solo si no se incluyen categorias}

\item{api_key}{elemento caracter con la google api para conectar la funcion con gemini}

\item{temperature}{numeric}

\item{cont}{elemento character para contextualizar los datos}

\item{cat_temperature}{numeric}
}
\value{
list un vector de caracteres con la classificacion, un elemento lsita con las categorias, y los datos originales
}
\description{
Està funcion analiza los datos a traves de Gemini, i genera una lista, un vector con las categorias que considere oportuna, i un elemento de texto donde encontramos una descripcion de cada categoria, para entender mejor que abarca.
}
\examples{
enq <- system.file(
"extdata",
"enquesta_satisfaccio.txt",
package = "IAIAdata"
)
api_k <- system.file(
"extdata",
"api_key.txt",
package = "IAIAdata"
)

datos <- readLines(enq)
api <- readLines(api_k)
class <- classifai(dt = datos,
                   api_key = api,
                   sol_tem = "Explicar las inquietudes de los pacientes")
}
