% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifai_bin.R
\name{classifai_bin}
\alias{classifai_bin}
\title{Selecciona categorias para classificar los datos}
\usage{
classifai_bin(dt, cat, api_key, temperature = 0.05)
}
\arguments{
\item{dt}{Character vector.}

\item{api_key}{character}

\item{temperature}{numeric}

\item{cont}{character}
}
\value{
data.frame con una columna por categoria indicando para cada elemento de la entrada si petenece o no a esa categoria
}
\description{
Està funcion analiza los datos a traves de Gemini, i genera una lista, un vector con las categorias que considere oportuna, i un elemento de texto donde encontramos una descripcion de cada categoria, para entender mejor que abarca.
}
\examples{
enq <- system.file(
"extdata",
"enquesta_satisfaccio.txt",
package = "IAIAdata"
)
api_k <- system.file(
"extdata",
"api_key.txt",
package = "IAIAdata"
)
datos <- readLines(enq)
api <- readLines(api_k)
classificacio_binaria_resultat <- classifai_bin(dt = datos, cat= cat, api_key = api)
}
