% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_cat.R
\name{select_cat}
\alias{select_cat}
\title{Selecciona categorias para classificar los datos}
\usage{
select_cat(dt, nmax_cat = "", sol = "", api_key, cont = "", temperature = 0.5)
}
\arguments{
\item{dt}{Character vector.}

\item{nmax_cat}{numeric}

\item{sol}{character}

\item{api_key}{character}

\item{cont}{character}

\item{temperature}{numeric}
}
\value{
list un vector de caracteres y un elementto caracte
datos <- system.file(
"extdata",
"enquesta_satisfaccio.txt",
package = "IAIAdata"
)
api <- system.file(
"extdata",
"api_key.txt",
package = "IAIAdata"
)
}
\description{
Està funcion analiza los datos a traves de Gemini, i genera una lista, un vector con las categorias que considere oportuna, i un elemento de texto donde encontramos una descripcion de cada categoria, para entender mejor que abarca.
}
\examples{
class <- select_cat(dt = dades,
                    api_key = api,
                    sol = "Explicar las inquietudes de los pacientes")
}
